(Here's hoping this takes) Added global_setup.h to declare init_world function
Moved variable declarations out of the interpreter loop for small speed boost.

Changed end opcode to now be opcode #0. A function is provided for it, but
isn't ever executed.

When the interpreter sees an opcode 0 in the stream it means the end
of stream. Things were coring on Cygwin otherwise. (I think I was
dereferencing a zero, which is bad)