This patch does two only somewhat related things.
Parrot/Opcode.pm:  Change the hard-wired assumption that sizeof(nv) =
2*sizeof(iv) to a PConfig-type computed value.

Parrot/Test.pm:  Change it to run with the perl that was used to
generate this version of Parrot.  This is useful if, for example, you
want to use bleadperl to build and test parrot (e.g. to try out various
64-bit options) but you still have a 5.6.x (or 5.005_xx) installed as
the default 'perl' on your system.  Since perl-5.7.x is installed
'version-only', a plain 'perl' is quite likely to pick up the older perl
in such cases. Such an older perl might well not understand pack('q')
used in the assembler.

Courtesy of: Andy Dougherty <doughera@lafayette.edu>