Welcome to the Just-in-Time Predereferencing Patch!
This patch implements a "predereferencing" mode for parrot. That means that
a lot of the dereferencing you see in DO_OP and in the opfuncs is done up
front (really, JIT) rather than as and each time we execute each op in the
byte code stream.

This works by creating an array of void pointers with the same number of
elements as the byte code array. This array starts out life zero-initialized.
Each time we go to execute an op, we check first to see if we have a NULL
pointer at *cur_opcode. If so, we know we haven't done our predereferencing
yet (this is the JIT part), so we call the prederef() function with sufficient
args that it can go look up opfunc pointers and op arg types and put the opfunc
and arg pointers at cur_opcode[0] .. cur_opcode[N - 1], where N is the number
of args (counding the op itself as one arg).

We assume that sizeof(INTVAL) <= sizeof(void *), so we can continue to have
inline INTVAL constants.

I've been getting about a 22% speed increase over vanila code path when I
compile with -O0. Advantage decreases at -O9 here. I tested with mops.pbc,
and also by temporarily putting the -P flag in Parrot/Test.pm and running
all the tests. So, this code is as correct as the test suite can detect.

Here's how it plays out in the code:

  * New script ops2c-prederef.pl

    Copy-paste code reuse from ops2c.pl. Modified to expect the args to give
    a void ** cur_opcode and to return a void **.

  * Modified ops2c.pl

      * opfuncs are now static. Who needs to call them by name from outside,
        anyway? They are accessed via the opfunc table.

      * Some MACRO stuff to make things work whether we prederef or not. It
        may be possible to get rid of these some day if other folks agree
        that we should be using CODE object pointers + offsets as our PC
        rather than raw pointers to either opcode_t's or (void *)'s
        (depending on prederef mode or no).

  * core.ops modified to be usable to build both kinds of opfuncs. For now,
    this means:

      * calls to push_generic_entry(... cur_opcode ...) were changed to
        push_generic_entry(... CUR_OPCODE ...), where CUR_OPCODE is a
        macro that does the right thing in both kinds of ops.c files.

        This is applicable to the bsr op. If we treated destinations as
        (Code PMC + local offset) rather than as pointer-to-memory-somewhere,
        this wouldn't be a big deal. We may need to do that at some point
        anyway, so that when you jump around and return not only do you have
        the right PC (pointer, today), but you also have the thread/interpreter
        knowing what chunk of bytecode it is in at all times (assuming we'll
        allow reflection, as in ops that query or affect the current CODE PMC).

      * runinterp was modified. Instead of copying the top-level PackFile struct
        and blasting its byte_code pointer, we simply call runops on the new
        interp, passing in the pointer to the current interp's code and the
        offset into that code at which to start (this last arg used to be a
        pointer PC, now its an offset PC).

  * include/parrot/op.h

      * new typedef for prederef opfuncs (they return void **)

  * include/parrot/interp_guts.h

      * New DO_OP_PREDEREF() macro described above.

  * include/parrot/interpreter.h

      * New prototype of runops() -- see below.

      * resume_addr field changed to resume_flag and resume_offset pair of fields.
        We might be resuming from interp using prederef to one not or vice-versa,
        so pointers between the two don't play nice together. Offsets are beautiful.

      * New PARROT_PREDEREF_FLAG constant

  * interpreter.c

      * include "parrot/oplib/core_ops_prederef.h"

      * prederef() function described above.

      * runops_prederef() runops variant. Uses DO_OP_PREDEREF() macro as its core.

      * runops() modified to take size_t CODE-relative offset PC rather than
        pointer PC. This made other things workable and doesn't seem to have lost
        anything important (IMO).

  * test_main.c

      * New flag '-P' to enable predereferencing.

  * Makefile modified to build core_ops_prederef.c from core.ops via
    ops2c-prederef.pl.

    BTW, also added missing $(INC)/interp_guts.h to H_FILES.

    Added core_ops_prederef$(O) to O_FILES