It adds a prompt to Configure asking for a list of opcode files to use.  The default is everything available except for obscure.ops.
In addition, it makes it so that if your answer to a question starts
with '+', it will concatenate it to the default.  For example:

What C compiler do you want to use? [cl] +HAHAHA!!!

will result in $PConfig{cc} eq "cl HAHAHA!!!".  While that example is
not terribly useful, it is quite useful with things like command-line
flags and the like.

Finally, it contains a couple tweaks so that my syntax-highlighting
editor recognizes some heredocs correctly.  :^)

Courtesy of "Brent Dax" <brentdax@cpan.org>