From:    Michel J Lambert <perl6@jall.org>
I'm sorry, this new weapon is going to give a quick advantage to the fools
side, but luckily, should help the fool-proofers in the long run.

Below patch should be safe to apply, as it does nothing. Turn on GC_DEBUG
if you want to see the hell that ensues.