Added macros for working with the interpreter->flags structure, following the _SET/_CLEAR/_TEST convention from pdd07.
Changed all the things which were directly accessing ->flags to use these
macros instead.  This gave me one place to put a hint comment to lclint,
which was complaining about bitwise operations on a non-unsigned variable
(apparently the enum type is signed by default).  I've set it to ignore this
by adding a /*@i1*/ to the appropriate lines.