Okay, finally adding the new assembler. Since this still is somewhat primitive it's not considered the "main" assembler yet. Especially since macros et al need to be added in.
To build this, use the 'make newasm' target. There probably will be some confusion over the PackFile vs. Packfile names, but that can't be helped for the moment. I'll take some time over the next few days to make sure that all of the test files &c are moved over to the new assembler, hopefully in some rational piecewise fashion.

This -does- do some work in lib/Parrot that may not be cleaned up, but that can be added in later. At least this will give us an opportunity finally to experiment with keyed aggregates and get those commands working.