-Implement PMC inc/dec functions and add ops to engine.  'inc P0, 5' is faster than 'set P0, P1, 5' which uses a vtable. -Corrected 'inc Nx, Ny' to 'inc Nx, Iy' as per the PDD -Added missing 'add Nx, Ny, Iz' op
Update mops_p.pasm to use dec op, however, the intent was
not to optimize mops_p benchmark, but to optimize for a very
common operation, which mops_p just happens to do. We now need
to re-evaluate mops_p and how to accurately represent a cross-section
of operations and not a specific case.