Peter's latest patch also included changes to the logic in terms of what to do when we don't have any available memory left in the pool. Patch is below.
Results are:
                before      after
gc_alloc_new.pbc        4.155999    0.18
gc_alloc_reuse.pbc      16.574      10.144998
gc_generations.pbc      4.025       4.175998
gc_header_new.pbc       3.686       4.286
gc_header_reuse.pbc     5.577999    4.446001
gc_waves_headers.pbc        3.815002    3.865
gc_waves_sizeable_data.pbc  8.383002    0
gc_waves_sizeable_headers.pbc   5.668       5.768

The '0' result comes from parrot crashing. This is not due to this patch,
but rather to GC bugs in parrot that this particular patch exposes.
This also causes stacks test 29/29 to crash for the same reasons. Since
it was included as part of peter's main patch which included a 'cycles'
solution to the GC problems, I don't believe the entire patch caused
parrot to crash.

gc_alloc_new seems to have improved a *lot*. This is because
gc_alloc_new allocates a lot of memory using the same headers. It
doesn't tear through headers quickly enough to trigger any dod runs on
its own, so these headers stay live and allocate tons of memory that gets
continually copied between generations.

The logic it uses is thus: If there's enough that's reclaimable, then we
compact to reclaim it. Otherwise, if there's not enough to reclaim, then
it's a hint that perhaps we need to run a DOD to reclaim some more, and so
it does.

Other results include a gc_alloc_reuse and gc_header_reuse improvement due
to the same reasons (the dod calls we wouldn't otherwise get). We get a
performance drop on gc_header_new due to the fact that we are allocating
lots and lots of headers. Performing dod runs doesn't help here, and only
serves to slow it down.


This patch seems to improve more than it deproves, so it should probably
go in, once Parrot is fixed to remove this GC bug (or perhaps before, to
ensure it gets fixed. ;) We should revisit it at some point to see if
there is a way to dynamically determine what kind of behavior our program
is doing, and avoid dod runs when allocating if we're in a massive
allocation stage of the program's life (program startups, calculation
setup, etc).

Courtesy of: Mike Lambert <perl6@jall.org>