* win32 can flush it's file buffers (FlushFileBuffers()) * SetFilePointer knows about whence, win32 constants (values, not names) are the same as in linux.
remarks:
FlushFileBuffers doesn't work for console handles, ms help file says:

Windows NT: The function fails if hFile is a handle to console output. That is because console output is not buffered.
The function returns FALSE, and GetLastError returns ERROR_INVALID_HANDLE.

Windows 9x: The function does nothing if hFile is a handle to console output. That is because console output is not buffered.
The function returns TRUE, but it does nothing.

well done microsoft...but it doesn't sound like a big deal. so calling the function should be just fine.

(Courtesy of Daniel Ritz <daniel_ritz@gmx.net>)