This patch puts the Perl6 Compiler in a much better state for win32 platforms.
Basically, it makes the Parrot::Test module a little bit more modular, so that non-Parrot tests can be performed using it's cross-platform abilities. If you're interested, all of the logic occurs on the last executable line of Parrot::Test and P6C::TestCompiler (which uses Parrot::Test now).

Currently, P6C now passes about 50% of the tests on Win32. Sean says that this is normal behavior for a fresh CVS checkout for now. (His local version has other patches which he is unsure about committing.)