Two changes:
a) minor doc fix in core.ops:
mem_allocs_since_last_collect is the number of new blocks allocated,
not the total memory allocated.

b) fix for the following problem:
The title says it all really: the counter in the interpreter structure
that tracks recent header allocations is initialized to 0 when the
interpreter is set up, but isn't incremented when headers are allocated.
Consequently, this:

interpinfo I1, 8
print I1

always prints zero.


Courtesy of Simon Glover.