I think that this patch is at good point to pause and take stock. I believe it JITs just about every integer op (including some i386 isn't JITting yet) OK, it doesn't JIT the logical xor op, but that one scares me, and I'm unsure how useful it is.
I've not done the floating point ops (or anything else) partly because I
don't have a good reference for the format of the floating point instructions.
[However, it's not that hard, as I have source code to both point emulators
supplied with ARM Linux, so I can see the decode code :-)]
But more because I'm not sure that it will give such a speed it.

Courtesy of: Nicholas Clark <nick@unfortu.net>