Here's most of the integer and floating-point arithmetic ops.  I didn't extend the register allocator to float regs, so they all just load and re-store their operands.  They pass all the tests in t/op/number that don't involve bsr or jsr.
Note that integer (not floating point!) sub is backwards for some reason
on the ppc.  I fixed this by reversing the way it's #defined in
jit_emit.h.

The jit still has some issues with branches, which cause it to fail a
number of tests.  I'd like to figure out why this is happening, but
probably not tonight.  It may have been a problem that was there before,
or it may be something I introduced.

Courtesy of: Sean O'Rourke <sorourke@cs.ucsd.edu>