A start on the python grammar...
Also, converter.pl is the tool I used to automatically convert the python.bnf file (epsilon a few bugs) to python.prd, which should work with Parse::RecDescent. The actual output in python.prd needed some minor editing, but one or two quirks can be fixed within the converter script.

Note, please, that converter.pl's grammar is horrendously ugly, and I'll be moving some of the terminals around to clean up the actual code.

The parser is also, right now, rather Python-specific. There's a weird "f"..."g" notation which only appears in Python BNF files, and the <this is a comment> notation. Also, I probably need to handle entity codes, but that's later on.