I discovered 'make languages' yesterday.  The enclosed patch cleans up a lot of small nits I found in the build process.  In a number of cases, the Makefiles were running perl scripts as         ./script rather than as         $(PERL) script
A few other places called a plain 'perl' instead of $(PERL).

Second, Configure.pl was putting the wrong flags in to build a shared
library.  (Or, more precisely, it was apparently unconditionally using
flags that work for GNU binutils.)  I left ld_shared_flags empty because
I don't know what is supposed to go there, but the value Configure.pl
used to use is definitely not right for Solaris's linker.

(Courtesy of Andy Dougherty.)