The use of ar rcs to build a library and reconstruct the symbol table is non-portable.  (Mac OS X, for example, doesn't support the 's' option to ar.)
A more portable -- but still Unix-specific -- approach is to run ar to
build the archive and then run ranlib.  For many many modern systems,
ranlib is not needed.  For those, the perl5 $Config{ranlib} entry ought to
be a simple ':' (i.e. a shell true command).

The following patch changes the main makefile to use ar and ranlib, just
as perl5 has successfully done for years.

(Courtesy of Andy Dougherty.)