In tracking down a gc bug, I realized that the current throwaway implementation of the print op could be replaced with a faster throwaway implementation that avoids doing a string_to_cstring.
Note that both the original and new implementations are still buggy
with respect to supporting different encodings. I don't know if
printf("%s") is any better than fwrite in terms of at least vaguely
paying attention to your locale or whatever. If so, don't apply it.

(Courtesy of Steve Fink)