Added a new core data structure and a corresponding PMC. It's really a dequeue (double-ended queue), meaning that push, pop, shift, and unshift are all constant-time operations, but arbitrary indexed lookup is theoretically linear-time. (It's linear in the number of 256-element chunks, so in practice it's usually fast.) I called it 'IntList' because IntDequeue is hard to type, fewer people know what it means, and list is just as accurate if a bit less specific -- it is a doubly-linked list with constant time access to the beginning and end, with the optimization that interior nodes are stored in compact arrays.
It is implemented as a modification of rxstack.c. It is still a
circular, doubly-linked list of chunks containing a fairly large array
of integers, only now the first chunk can be partial as well as the
last, in order to efficiently support shift/unshift. The length is
cached in the first node. Also, these chunks are now Buffer
subclasses, so they are completely managed by GC (rxstack uses
mem_sys_allocate).