# New Ticket Created by  Jason Gloudon # Please include the string:  [perl #18520] # in the subject line of all future correspondence about this issue. # <URL: http://rt.perl.org/rt2/Ticket/Display.html?id=18520 >
This patch just adds a new control flow flag to op.h and extracts the
additional information from op definitions needed to identify ops that cause
interpreter retarts. I've changed the output to the opcode C file so that it
uses the enum values in OpsFile.pm instead of hard coding their integer values.

The current approach to handling restart in the JIT checks for restart after
calling every external op function that can adjust the flow of control. The
normal flow control ops cannot cause a restart, so those checks are mostly
unnecessary.

With the additional information the unnecessary checks can be removed for ops
that cannot cause restarts.