Alright, here's another string_str_index patch.
This one screams at 2.5 times the speed of Perl 5 for the single byte
encoding.  Multibytes are implemented with a stupid, slow algorithm.

If anyone knows of a way to implement the Boyer-Moore algorithm with
multi byte encodings, please tell me or do so.  (The only way I can
think of is with a hashtable, which would probably be slower than the
brute force method implemented here.)

This is still only my second patch, so considerations, critiques,
etc. are most welcome.

 - Luke Palmer