Some perls, such as Perl 5.6.1, configured for long doubles don't work *or* fail correctly when packing "D".  The assembler thinks they are being packed in - for x86 - 12 bytes, when they're really packed in 8. So the constants section is being packed with 8 byte numbers, while the constant header reflects lengths assuming 12 byte numbers.
To make matters worse, regardless of what size bytes the floating types
are, the packfile header reports them as doubles.  (Assuming I
understand the purpose of the header's floattype field.)

This patch hacks the assembler to fall back to assembling floats as
doubles in the case that Perl can't pack long doubles, or that long
doubles result in a float not covered under the floattype header
section.  (It also fixes that header to report when 12 byte numbers
are indeed packed.)  The fix is *solely* to the assembler.  There's
little reason to not allow parrot to (theoretically) process long
doubles just because it can't assemble them.

This patch, when couple with the previous several, causes Parrot-x86,
configured for long long INTVALs, long double FLOATVALs, and long long
opcodes, to pass all non-skipped tests, when based on Perl 5.6.1.

Generic huge INTVAL, FLOATVAL, and opcode configurations are still
undoubtedly broken, as well as various combinations of the above
configurations.

In particular, long double FLOATVALs when based on Perl 5.8.0, which
should exercise the 12 byte format, remains untested.  I'm going to run
through the remaining scenarios and verify what works and what doesn't.

Courtesy of Bryan Warnock <bwarnock@raba.com>