Half-step towards a usable module system for Jako.
  * Changed the property syntax from {attr[=value],...} to :attr[=value]...

    Including changes to example programs to use new syntax.

  * Added a new construct: module

    Allows the grouping of sub declarations / definitions with a namespace.

  * Changed examples/nci.jako to use an embedded copy of the Curses module
    to reference the functions in the curses dynamically loadable library.

NOTE: Before modules are truly useful, they will need to be loadable from
separate files early on at compile time, probably with #line directives to
keep things straight. For now, modules are great for grouping, but they
have to be inside your single source file. And, all symbols in the module
are exported to the parent namespace automatically.

NOTE: The :: "delimiter" in namespaces is mapped to "__" for the benefit
of IMCC, since it does not like "::" in identifiers.