I've added a way to add or remove options on a per-c-file basis.
Right now all it does is make sure that tsq.c won't be affected by
Configure.pl --optimize, and that the amounts of silly warnings is
reduced (by removing -Wformat-nonliteral on files we know use that
feature..), as well as removing a couple of sparc-specific(?) VERY
noisy warnings. On top of this, the Makefile rewrite hack for imcc
has been eliminated.

Courtesy of Lars Balker Rasmussen <lars@balker.org>