currently Configure.pl insists on passing a -mno-accumulate-outgoing-args option to GCC 3.x.
a trustworthy source (Nicholas :-) told me that this should only be
available on x86 and x86-64, and in fact my ia64 complains that the
option is invalid.

I've patched /config/auto/gcc.pl to check if $Config{archname} =~
/86/, and in case it doesn't, remove the offending option.

I'm not sure this is the correct approach, but it shouldn't affect
negatively other systems/versions, and for sure it works on my ia64.

Courtesy of Aldo Calpini <dada.list@alos.it>