#24880 Here is a patch to add a few more jitted ops to core.jit for ppc (noop, neg_i, neg_n, abs_n, abs_i).
It also modifies build_tools/jit2h.pl slightly to allow for templates
based on other templates (the patch just moves a block of code already
in the script). This makes some things a bit tidier, but if it's
problematic then I can rework the patch to core.jit to work without it.

Courtesy of Jeff Clites