[PATCH] more oo*.* benchmarks
My Parrot, Python, or Ruby-fu are not as strong as they should be
(caveat applicator), but here goes nothing: I added some simple oo
benchmarks for 'getters' and 'setters'.  In the attached .tgz (destined
for examples/benchmarks) the included oon.txt explains what the heck are
all the different files, and why the oo[56].pasm are missing.  I also
tweaked some of the existing files (oo[12].{py,pasm}) so that the
benchmarks go through the same range (1..x00000).

Courtesy of Jarkko Hietaniemi <jhi@iki.fi>