[perl #29315] [PATCH] Win32 Intel C++ Compiler support
  Some changes so that "perl Configure.pl --cc=icl" correctly uses
the Intel C++ Compiler under Windows:

  config/gen/icu.pl is just updated to think that icl and icl.exe are
MSVC; it'll do the ICU compilation with the MS compiler, but the
resultant library is compatible with the Intel compiler.

  config/init/hints/mswin32.pl handles icl similarly to cl, but with
a different cc ('icl') and ld ('xilink') and ar ('xilib'), and also
adds libircmt.lib to the library list.

icl doesn't like the emptiness of:
    const struct {
[ ... ]
    } _temp_mmd_init[] = {

  With those changes, it all compiles and passes the tests (and still
works in MSVC).
  [All done with version 8.0.048 of the Intel compiler, and MSVC 7.1]

Courtesy of Philip Taylor <philip@zaynar.demon.co.uk>