[perl #29333] NCI Nested Struct Access Broken?
The attached test patch demonstrates what I think is a problem.  Running
gdb and looking through unmanagedstruct.pmc makes me wonder if the
PMC_val of the nested struct is ever actually assigned to the enclosing
struct.

Courtesy of Chromatic <chromatic@wgz.org>

The code didn't handle the case where we are creating a nested struct,
which is refered to by a pointer, like the given example:

  int bpp = surface->format->BytesPerPixel;

Fixed now.