[perl #29257] Build libnci.so by default
'libnci.so' is used for testing the native call interface. However I noticed
that the tests in t/pmc/nci.t were skipped, because 'libnci.so' wasn't built
by default. The attached patch adds that library to the target 'all'.

Previously 'libnci.so' was built in the 'parrot' directory. A better
location might be 'runtime/parrot/dynext'.

Courtesy of Bernhard Schmalhofer <Bernhard.Schmalhofer@biomax.de>