The following patch enables the user to control which compilers get chosen by ICU's configure, and tries to chose sensible defaults if the user does nothing.  By default, it runs ICU's configure with
	CC=$cc ./configure [options]

where $cc is the C compiler used to build parrot.  This is useful if ICU's
default choice of compiler isn't appropriate.

As an option, this patch allows you to specify the C++ compiler as well,
with something like

	perl Configure.pl --cxx=CC

Courtesy of Andy Dougherty <doughera@lafayette.edu>

--

Use fib 28 for benchmarks (leo)