[perl #30878] [PATCH] tru64 aio detection
-laio needed.

[perl #30879] [PATCH] IRIX offsetof

Some compilers of IRIX seem to have a quite broken definition of
offsetof.  Hand-roll our own if such a compiler is detected.

[perl #30880] [PATCH] const* vs non compiler silencing

[perl #30882] [PATCH] IRIX 64-bit compilation

If the compiler is 64-bit (cc -64 has been used) also the cxx, ld, and
link need to be set accordingly.

All
Courtesy of Jarkko Hietaniemi <jhi@iki.fi>