resizablefloatarray was leaking memory at pop_float (the float buffer was reallocated but not reattached to the pmc in case the buffer moved).
While fixing that (by using set_integer_native(), just like push_float)
also implemented shrinking: if the new (used) size falls below one half
of the old (allocated) size, shrink the buffer to one and half times the
new size (up to three thirds of the old size, but using the new size avoids
truncation to zero at new size of one)  We do not shrink straight down
to the new size in anticipation of a future mixture of pushes and pops,
with this hysteresis we avoid frequent resizing at size boundaries.

This scheme could probably be used (especially for the pop op) in the
other resizable*array.pmc classes.