The attached patch fixes the solaris hints file to force the use of 'c++' for linking if Configure.pl finds gcc.  Without this patch, it links with gcc which fails since it apparently can't find some of the c++ symbols from icu.
This patch also changes the order of the Configure.pl tests so that
the gcc detection step is run before the hints step.  This has the
side effect of fixing the test that was already in the solaris hints
file.  As near as I can tell, this doesn't break anything, at least on
solaris.

Courtesy of coneill@oneill.net