[perl #31424] [RESOLVED] PATCH: Fix for parrot linking issue on Solaris 8
Attached is a patch that should fix the solaris link issues without
breaking every other platform, hopefully. ;)

This adds support for setting triggers on specific config variables.
The basic idea is that you can register a named callback for any
specific config variable.  When that variable is set, all the
callbacks registrered for that variable will be called to notify them
that the new setting is available.  I added subs in
config/init/data.pl to for adding triggers, getting them and deleting
them.

I also fixed the solaris hints file to use the new trigger code.  It
sets a trigger on 'gcc_version' and when that gets set later on in the
config, if it's defined, then the link command will be set to 'g++'.

Courtesy of Clayton O'Neill <coneill@oneill.net>