[perl #31962] [PATCH] synchronous NCI callbacks
A callback sub is marked as 'sychchronous' by setting the property
'_synchronous' of the user_data to a true PMC. A neater solutin would be
automatic detection of safe synchronous execution. When the external library
calls the callback, then no callback event is put into the event queue, but
the sub is executed immediately.

A test is included in t/pmc/nci.t .

In src/nci.c the variable names 'user_data' and 'external_data' are now
consistently used.

Courtesy of Bernhard Schmalhofer <Bernhard.Schmalhofer@biomax.de>