[perl #32466] [PATCH] Parrot m4 0.0.10 and "eval" changes
this patch brings Parrot m4 to terms with recent "eval" changes. The compile
function of the 'eval' compiler now returns an Eval PMC. The m4 macro "eval"
is a simple interpreter of integer arithmetic expressions.

The 'eval' compiler returns a bytecode segment without a constant table. The
'destroy' of the Eval PMC needs to handle that.

The compiler for 'japh16' now emits a 'invoke P1' instead of an 'end'.

Courtesy of Bernhard Schmalhofer <Bernhard.Schmalhofer@biomax.de>