[perl #32699] [PATCH] benchmark tests
This patch will allow all the *.imc and *.pasm benchmarks to be
tested by running make testbench.

There are currently two benchmarks ( hash-utf8.pasm and b6t.imc )
that fail.  The hash-utf8 one looks like broken code but the b6t one
seems to be something with exceptions.  The expected output for both
of those is incorrect.

For the rest I gleaned the expected outputs from running the benches
themselves so if the output is currently wrong then the tests will
be verifying that:)  Any help in realizing better/proper output
would be greatly appreciated.

Courtesy of Justin DeVuyst <justin@devuyst.com>