RT#31859, Plain ole Hash
I assume that the Hash PMC should also be used by PIR code in
'runtime/parrot/library'. In that case returning PMCNULL is not nice, as
the returned PMC can't even be queried for it's type.
So for now a None PMC is returned for non-existent keys.
However this is not consistent with e.q. FixedPMCArray, which does
return PMCNULL. ResizablePMCArray is even more inconsistent and seems to
return an Undef PMC.

New in this patch:
- removed the NCI method 'fromkeys' in the Hash and PerlHash PMC.
- PerlHash now gets all of it's C-Code from 'hash.pmc'
- removed the static 'undef' in classes/hash.c. I hope that this won't
harm performance.
- Added get_integer() and get_number() to the None PMC
- Added a test which queries a non-existent element from a FixedPMCArray
- Got rid of PerlHashes in 't/', 'imcc/t/', and 'runtime/parrot/library'
- Added a couple of tests for the Hash PMC

Courtesy of Bernhard Schmalhofer <Bernhard.Schmalhofer@biomax.de>