[perl #33801] [PATCH] GDBMHash - a dynamic PMC that binds to GNU dbm
I have been working on a dynamic PMC that calls into 'gdbm'. 'gdbm' is a
file based database, that provides dictionary lookup.

There are no fancy features yet. INVAL, FLOATVAL, PMC and STRINGS are
converted to char arrays and are written to the database file. The
conversions are not unlike for the FixedPMCArray PMC.
The keys are also converted to strings, just like in the Hash and PerlHash
PMC.
Fetched values are either STRINGs or String PMCs.

The availability of 'libgdbm.so' is checked during the Parrot configuration,
which is not nice.

Courtesy of Bernhard Schmalhofer <Bernhard.Schmalhofer@biomax.de>