[perl #34039] Build failure on freebsd 4.9-RELEASE (mpz_*)
The point is that the so-called mpz_fits_* functions were defined back
in 2004-04-07 according to the GMP ChangeLog file and that corresponds
to version 3.0. In a FreeBSD 4.10 box I have found a "gmp.h" at
"/usr/include" which corresponds to version 2.0 and does not have
mpz_fits_* functions.

... In config/auto/gmp/gmp.in the
problematic function is used and its answer reaches STDOUT too. Then
config/auto/gmp.pl tests for the extended expected output.

Courtesy of Adriano Ferreira <a.r.ferreira@gmail.com>