HP-UX build notes
I've retested on Linux and cygwin (although cygwin has problems with
dynclasses), and all's well.

A summary of the changes:

  * Two extra flags are made available (cc_shared, which was already
extracted from Perl but not used, and ccdlflags, which probably needs a
better name!), and added to the makefile etc.

  * I've removed -g from ld_debug, hopefully not upsetting anyone.
Platforms that need it should probably add it as I would have thought
that they are in the minority; can a Solaris person confirm that it's ok
for them.

  * Turn off aligned_funcptr for HP-UX; is that ok how I've done it?

With this patch HP-UX builds 'out of the box' for me (as hopefully will
other future platforms), but we still need a viable solution to get
execute permissions onto libraries in runtime/parrot/dynext/*.

The signal test fails due to the shell, and two of the nci tests hangs.

Courtesy of Nick Glencross <nickg@glencros.demon.co.uk>