[perl #36223] [PATCH]Loop Preheader
This patch adds a new function, natural_preheader(), to cfg.c.  This
function finds and returns the preheader of a loop, if one exists.
The preheader of a loop is the only predecessor of the loop's header
outside of the loop, and the header is the only successor of the
preheader.  Therefore, the preheader's code always executes
immediately before entering the loop body.  The preheader is useful in
the Loop-Invariant Code Motion optimization.

Courtesy of Curtis Rawls <cgrawls@gmail.com>