[PATCH] [perl #36286] [PATCH] Turn off optimization on another large file.
My system couldn't handle an optimized compile of ops/core_ops_switch.c.
Ideally, there should be some easy way to put such things in hints files,
and also some way to override it on a per-file basis after running
Configure, e.g. with something like make OPTIMZE="" .  But for now, this
will do for me.

Courtesy of Andy Dougherty <doughera@lafayette.edu>

[PATCH] [perl #36269] Alignment problems with doubles on SPARC.

The following patch fixes the core dump for me.  The problem appears to be
in <stacks.h>, where the 'data' segment is declared as a (void *), but is
then assumed to be aligned suitably for holding a double.  I fixed that
with a union.  The compiler then automatically assures that the elements
are aligned.  I still need to go through places where the data element is
assigned to various things through casts and check whether any additional
members to the union ought to be added.

While hunting for a solution, I saw a lot of places in the code that
appear to be worried about alignment.  In many cases it was not clear to
me what needs to be aligned with what and why.  Is it for alignment of
doubles?  Then we should use a double type where appropriate.  Is it for
alignment of one structure with another (like perl's sv/av/hv/
structures)?  If so, that should be documented carefully somewhere.  Is it
for alignment with blocks returned by memalign?  I didn't know, so I put
in questions in the comments.

[ stacks.h slightly modified to keep necessary alignment for x86 ]

Courtesy of Andy Dougherty <doughera@lafayette.edu>