[PATCH] Library Loading On Win32
Under Win32 it's common to install stuff...well...anywhere the person
installing it wants.  With this patch, we can create binary distributions
for Win32 that don't need Parrot to be installed in the place that "make
install" decides to put it.

Normally libraries are loaded by appending a prefix, stashed in the
executable at compile time, to some common library search paths.  With this
patch, if that doesn't work out for us then and we're on Win32 then we get
the path of the Parrot executable and use that as a basis for trying to find
libraries.

Courtesy of Jonathan Worthington <jonathan@jwcs.net>