[perl #37003] [PATCH] Change GCD and LCM opcode to use Binary GCD algorithm
Presently GCD/LCM use the Euclidlean algorithm to calculate the GCD,
the attached path replaces it with the Binary GCD algorithm. The
primary advantage of the Binary GCD algorithm is that it doesn't use
division like the Euclidlean algorithm (it uses bitwise shifts
instead) and so is more efficient. On a PC it is roughly 25% faster,
on embedded platforms which don't have processor based division
support it can be around 80% faster.

Courtesy of Imran Ghory <imranghory@gmail.com>