[perl #37609] Add return type to new_callback signature
As hinted in an earlier email, I'd like to propose the addition of a
return type to the new_callback signature. This will add consistency,
and prevent breakage when we support other NCI backends.

I've also made a change to the 'unhandled signature' which didn't look
like it was printing the signature correctly.

All NCI tests pass.

Courtesy of Nick Glencross <nickg@glencros.demon.co.uk>