[perl #38176] [PATCH] Fix string-to-integer conversion with multiple leading spaces
Until r10938, string-to-integer conversion was possible even when the
string contained more than one space before the start of the integer.

This patch modifies src/string.t to make this work again.

This patch also modifies t/op/string.t to add a test for this case, and
for other corner-cases of string-to-integer conversion.

Courtesy of Roger Browne <roger@eiffel.demon.co.uk>