Lua :
this patch changes the behaviour of the LuaTable PMC.
changes have to do with the reference semantics that PMCs have, and LuaTables should not have.

Setting and getting a value now will set (and get) a copy of the value.

Clone() returns a reference to itself, as clone() should be used for register to register copying (due to the by-value semantics of Lua values)

Courtesy of Klaas-Jan Stol.