Added find_namespace(), find_sub(), and find_var() to NameSpace PMC with tests.
Note that find_var() doesn't do any checking on the *type* of the PMC it finds,
so it can just as easily return a NameSpace or a Sub PMC as anything else.
It's possible to add a type check to ensure that this doesn't happen, but it's
not clear from the PDD if that's necessary or useful.