t/perl/Parrot_IO.t test 51 ("modified since") has an intermittant failure. Despite sleeping for a second, the newly modified file's mtime is not guaranteed to be $time+1.  When the file's mtime is equal to $time, the test fails.
I tried everything I could think of.  It doesn't look like any rounding is
occurring in time()'s return value, and tweaking $| doesn't help.  It looks
like a weird issue with perl5 behavior.

Work around it by sleeping for 2 seconds, when necessary.  This reduces the
dependency on perl5's semantics, while still testing the correctness of Parrot
library code (which is the point of it all, of course.)