[src] Optimized Parrot_byte_index(), which was surprisingly expensive.  It's a little bit tricky because C strings can't have embedded NULL characters, so the naive use of the standard library's index() doesn't always work.  However, it's a good place to start as it tends to run much faster than the inchworm code here.
Another optimization is never paying the price of calling memcmp() if it's
obvious that it will fail.

There are larger algorithmic optimizations to make calling this code less
expensive (by calling it less frequently), but this was a low-hanging
optimization.  Amortized over the data of PGE's regexp tests, this function
runs about three times faster than the previous version.