[PMC] Fixed a DOD-related segfault in NameSpace PMC (RT #46513).
When re-using a freed PMC for a Namespace due to a clone(), the destination PMC
would not be initialized, which left old data in the PMC_pmc_data() slot.
Hash's clone() used pmc_new_noinit() so it could create the underlying hash
itself, passing a pointer to the Hash pointer to parrot_hash_clone().

Now parrot_hash_clone() merely takes a pointer to the destination hash (because
you can modify the hash if you have a pointer to it, not just a pointer to a
pointer to it), so Hash's clone() can call pmc_new_init() and Namespace's
init() can initialize the Namespace PMC's data appropriately and there's no old
data lying around to confuse Namespace's mark() which prevents the GC from
trying to do bad things with invalid data.