[config] Add the ability to convert PBC files to executables linked against libparrot.
The generated pbc_to_c.pir file takes a PBC file as input, creates a .c file
from it, and attempts to compile that to an executable.  This works for me and
should be reasonably cross-platform, as long as you have dynamic linking
working for your libparrot.

The next step is to run pbc_to_c.pir on itself, probably in the root Makefile,
so that we know it's working and we can run it on other programs with just
pbc_to_c.

(This *does* work on perl6.pbc successfully, but we have to figure out PBC
paths appropriately to be able to run it anywhere except the top-level Parrot
build directory.)