[PMC] Migrated NCI PMC to PDD 17 for attributes.
Using the type PMC * instead of void * for the function pointer is a nasty
hack, but PDD 17 doesn't (yet) support void * as an attribute.

For now this only requires some dodgy type casting, but if and when the PMC
parser handles mark() and destroy() automatically, treating a function pointer
as a PMC is only asking for trouble.