[PMC] Changed ParrotLibrary PMC creation so that they are now constant, as is their metadata.
When performing full global destruction, occasionally a ParrotLibrary PMC would
get destroyed before the dynpmc which had vtable entries from the shared
library contained by the ParrotLibrary.  Because ParrotLibrary closes shared
library handles in its own destroy() entry, the function pointers in the
dynpmc's vtable became invalid.  Thus anything which attempted to call
destroy() on a dynpmc caused a segfault, as the function went away.

Note that this was a heisenbug, as the order of operation mattered, and this
behavior depends on the particular ordering of PMCs with custom destroy actions
in the non-constant pools.

Now that ParrotLibrary PMCs are constant, they get destroyed only after all
non-constant PMCs.