[GC] Fixed PMC_EXT allocation/deallocation.
PMC_EXT structures are *not* isomorphic with the PObj structure.  Treating it
as such was an unnecessary hack that is now gone.  There is one smaller hack in
the revised version, but it's part of the free list pointer hack that's already
in the code, and it's much better than the troublesome casting and
bit-twiddling necessary to treat PMC_EXT headers and PObj headers somewhat
interchangeably.