[t] Fix an intermittant failure that's been around forever.
The "three alarms" test in t/dynoplibs/myops.t has a race condition.

When run on a machine with a high loadavg (I used loadavg=15 for my 
tests), the "main" subroutine doesn't even get around to queueing the 
later alarms before the former ones have already fired.  This results in 
the output statements occurring out of order, which fails the test.

Since the order of these things' execution is not guaranteed, 
particularly under high load when there are significant delays just 
queueing the things, the test shouldn't depend on their order.  Fix 
the test so that each callback sets 1 bit of a "flags" integer, all we 
care about is that the callbacks were actually called.