[tcl] http://code.google.com/p/partcl/issues/detail?id=59
Part of the eventual switch to PCT involves changing how we attempt to compile.
This reverts tcl to a pure runtime dispatch for each command; The ".tmt"
files we had been using to generate PIR to use in place of the runtime call
are now gone.

In most cases, just used the version that had been automatically generated by
the original system.