[GC] Made the concurrency scheduler a constant PMC.  This means that it never gets collected by the end-of-the-world GC run before all tasks and timers and events (which depend on it for their destruction) get collected and destroyed. See RT #60622 for one example of that.
Of course, as it's a constant PMC now, it always needs its live flag cleared
just before it gets marked as live, otherwise it won't have a chance to perform
its custom marking of the PObjs to which it refers.  This is a workaround, but
the real fix is to figure out safe and consistent order of destruction, and
that's a much larger patch.

I also tidied some related code.