[io] Fix Parrot_io_write_buffer()'s buffer rotation case
In the final block, this function filled the (nearly already full) buffer,
flushed it, and then put the rest of the data into the (now empty) buffer.
However, it didn't set the "buffer has data in it and needs to be flushed"
flag (which had been unset by Parrot_io_flush), and the extra data was
dropped.

Set the flag.  This fixes TT #296.