[IMCC] Allowed Unicode identifiers in named parameters, arguments, and return values, as TT #654 requests.
The resulting IMCC code could use some refactoring: create a new token which
represents a constant SymReg that is already a properly constructed STRING or
Unicode STRING.  As well, this new code needs memory profiling, as the expected lifespan of this memory isn't always clear.  Still, it fixes the immediate bug.