[PMC] Fixed segfaults in ImageIO PMC's destroy().
In certain (yet undiagnosed) cases, the "You can destroy the PackFile
attribute" flag is on but there's no PackFile attribute.  This makes C sad.
The culprit is r48447.

A better approach is to use PObj_custom_destroy_SET() only when it's okay to
destroy this attribute, but that fix will take longer.  In the meantime, this
works at the expense of a bit of extra work.