[OO] Removed PIR overriding of Object's get_class.
You're asking for strange and wonderful and bad behavior if you do this, so
nothing does it, so why pay the NQP-rx penalty for it?  When we get a more
flexible metaobject system, people can do the kinds of things they need here in
a different (better and safer) way.