[GC] Fixed unoptimized builds in stack tracing.
The STRING marking function has an assert for STRINGness that Buffers which get
marked don't need.  This whole system is a mess of false and foolish premature
isomorphism, but we can live with this for now (unless someone beats me to
cleaning up this approach before we replace this part of the GC).