[str] Encoding cleanup
There are several places where strings are created with the default
ASCII encoding and filled with binary data. This commit fixes hopefully
all of these cases and adds checks for ASCII data in Parrot_str_new_init
and STRING_iter_set_and_advance.

This commit also adds the "b" flag to FileHandle.open as a shortcut to
open a file in binary mode.